+++ packages/component/CHANGELOG.md
@@ -450,6 +450,70 @@
 - Updated dependencies
   - @orderly.network/types@0.1.0
 
+## 0.0.100
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/net@1.0.90
+  - @orderly.network/types@0.0.92
+  - @orderly.network/utils@0.0.78
+
+## 0.0.99
+
+### Patch Changes
+
+- fix bug
+
+## 0.0.98
+
+### Patch Changes
+
+- fix bug
+
+## 0.0.97
+
+### Patch Changes
+
+- for test
+
+## 0.0.96
+
+### Patch Changes
+
+- add op chain
+- Updated dependencies
+  - @orderly.network/types@0.0.91
+
+## 0.0.95
+
+### Patch Changes
+
+- add cross chain withdraw
+- Updated dependencies
+  - @orderly.network/types@0.0.90
+
+## 0.0.94
+
+### Patch Changes
+
+- fix bug
+- Updated dependencies
+  - @orderly.network/net@1.0.89
+  - @orderly.network/types@0.0.89
+  - @orderly.network/utils@0.0.77
+
+## 0.0.93
+
+### Patch Changes
+
+- fix bug
+- Updated dependencies
+  - @orderly.network/net@1.0.88
+  - @orderly.network/types@0.0.88
+  - @orderly.network/utils@0.0.76
+
 ## 0.0.92
 
 ### Patch Changes