+++ packages/tsconfig/CHANGELOG.md
@@ -1,5 +1,17 @@
 # tsconfig
 
+## 0.0.84
+
+### Patch Changes
+
+- update
+
+## 0.0.83
+
+### Patch Changes
+
+- fix bug
+
 ## 0.0.82
 
 ### Patch Changes